requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
python-dotenv
qrcode[pil]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
